refactor: rename motion to fancy
