+++ src/components/fancy/link/anim.ts
