+++ src/components/fancy/link/word.tsx
