+++ src/components/fancy/motion-trail.tsx
