+++ src/components/fancy/number-ticker.tsx
