+++ src/components/fancy/text-loop.tsx
