+++ src/components/fancy/text-reveal.tsx
