+++ src/components/fancy/variable-font-cursor-proximity.tsx
