+++ src/components/sections/header/compact/header.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React, { useState, useEffect } from 'react';
-import Link from '@/components/motion/link';
+import Link from '@/components/fancy/link';
 import { motion } from 'motion/react';
 import { CodeIcon, EllipsisIcon, MenuIcon, XIcon } from 'lucide-react';
 import ThemeToggle from '@/components/mode-toggle';
