+++ src/components/sections/projects/compact/projects.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import MotionWrap from '@/components/motion-wrap';
+import MotionWrap from '@/components/fancy-wrap';
 import ProjectCard from './project-card';
 
 import {
