+++ src/components/sections/projects/modern/projects.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import MotionWrap from '@/components/motion-wrap';
+import MotionWrap from '@/components/fancy-wrap';
 import ProjectCard from './project-card';
 
 import {
@@ -12,7 +12,7 @@ import {
 } from '@/components/ui/carousel';
 
 import { project } from '@/app/source';
-import TextReveal from '@/components/motion/text-reveal';
+import TextReveal from '@/components/fancy/text-reveal';
 
 function Projects() {
   const projects = [...project.getPages()].sort(
