+++ src/components/sections/technologies/modern/technologies.tsx
@@ -1,10 +1,10 @@
 'use client';
 import React, { useRef } from 'react';
-import MotionWrap from '@/components/motion-wrap';
+import MotionWrap from '@/components/fancy-wrap';
 import TechnologyCard from './technology-card';
 
 import { technologies } from '@/components/sections/technologies/config';
-import TextReveal from '@/components/motion/text-reveal';
+import TextReveal from '@/components/fancy/text-reveal';
 
 function Technologies() {
   const containerRef = useRef<HTMLDivElement>(null);
