+++ src/components/sections/testimonials/compact/testimonials.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import MotionWrap from '@/components/motion-wrap';
+import MotionWrap from '@/components/fancy-wrap';
 import {
   Carousel,
   CarouselContent,
